refactor(controllers): clean-up code
Replace general `@RequestMapping` annotations with specific annotation
`@GetMapping` reflecting HTTP method.

Remove unnecessary path variable definitions.

Rename package.
