--- Poseiden-skeleton/src/main/java/com/nnk/springboot/controllers/HomeController.java
@@ -1,23 +1,0 @@
-package com.nnk.springboot.controllers;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-public class HomeController
-{
-	@RequestMapping("/")
-	public String home(Model model)
-	{
-		return "home";
-	}
-
-	@RequestMapping("/admin/home")
-	public String adminHome(Model model)
-	{
-		return "redirect:/bidList/list";
-	}
-
-
-}
