+++ Poseiden-skeleton/src/main/java/com/pcs/tradingapp/controllers/BidListController.java
@@ -1,0 +1,54 @@
+package com.pcs.tradingapp.controllers;
+
+import jakarta.validation.Valid;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+
+import com.pcs.tradingapp.domain.BidList;
+
+@Controller
+public class BidListController {
+    // TODO: Inject Bid service
+
+    @GetMapping("/bidList/list")
+    public String home(Model model)
+    {
+        // TODO: call service find all bids to show to the view
+        return "bidList/list";
+    }
+
+    @GetMapping("/bidList/add")
+    public String addBidForm(BidList bid) {
+        return "bidList/add";
+    }
+
+    @PostMapping("/bidList/validate")
+    public String validate(@Valid BidList bid, BindingResult result, Model model) {
+        // TODO: check data valid and save to db, after saving return bid list
+        return "bidList/add";
+    }
+
+    @GetMapping("/bidList/update/{id}")
+    public String showUpdateForm(@PathVariable Integer id, Model model) {
+        // TODO: get Bid by Id and to model then show to the form
+        return "bidList/update";
+    }
+
+    @PostMapping("/bidList/update/{id}")
+    public String updateBid(@PathVariable Integer id, @Valid BidList bidList,
+                             BindingResult result, Model model) {
+        // TODO: check required fields, if valid call service to update Bid and return list Bid
+        return "redirect:/bidList/list";
+    }
+
+    @GetMapping("/bidList/delete/{id}")
+    public String deleteBid(@PathVariable Integer id, Model model) {
+        // TODO: Find Bid by Id and delete the bid, return to Bid list
+        return "redirect:/bidList/list";
+    }
+}
