+++ Poseiden-skeleton/src/main/java/com/pcs/tradingapp/controllers/HomeController.java
@@ -1,0 +1,21 @@
+package com.pcs.tradingapp.controllers;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class HomeController
+{
+	@GetMapping("/")
+	public String home(Model model)
+	{
+		return "home";
+	}
+
+	@GetMapping("/admin/home")
+	public String adminHome(Model model)
+	{
+		return "redirect:/bidList/list";
+	}
+}
